+++ src/components/formcomponents/date/date-month-input.tsx
@@ -8,13 +8,13 @@ import Select from '@helsenorge/designsystem-react/components/Select';
 
 import { useMinMaxDate } from './useMinMaxDate';
 import styles from '../../../styles/date-year-month.module.css';
-import { getId, isReadOnly, isRequired } from '../../../util';
+import { getId, isReadOnly } from '../../../util';
 import styles2 from '../common-styles.module.css';
 import { ReadOnly } from '../read-only/readOnly';
 
 import { QuestionnaireComponentItemProps } from '@/components/createQuestionnaire/GenerateQuestionnaireComponents';
 import { ReferoLabel } from '@/components/referoLabel/ReferoLabel';
-import { getErrorMessage } from '@/components/validation/rules';
+import { getErrorMessage, required } from '@/components/validation/rules';
 import { shouldValidate } from '@/components/validation/utils';
 import { useExternalRenderContext } from '@/context/externalRender/useExternalRender';
 import { useGetAnswer } from '@/hooks/useGetAnswer';
@@ -121,31 +121,25 @@ export const DateYearMonthInput = ({
   };
 
   const validationRulesYear: RegisterOptions<FieldValues, string> | undefined = {
-    required: {
-      value: isRequired(item),
-      message: resources?.yearmonth_field_required || '',
-    },
+    required: required({ item, resources, message: resources?.yearmonth_field_required }),
     validate: {
       validYear: value => {
-        return doesAnyFieldsHaveValue() ? validateYearDigits(getYearFromString(value), resources) : true;
+        return doesAnyFieldsHaveValue() ? validateYearDigits(getYearFromString(value), resources, item) : true;
       },
       validMinDate: value => {
         const monthValue = getValues(`${idWithLinkIdAndItemIndex}-yearmonth-month`);
-        return doesAnyFieldsHaveValue() ? validateYearMonthMin(minDateTime, getYearFromString(value), monthValue, resources) : true;
+        return doesAnyFieldsHaveValue() ? validateYearMonthMin(minDateTime, getYearFromString(value), monthValue, resources, item) : true;
       },
       validMaxDate: value => {
         const monthValue = getValues(`${idWithLinkIdAndItemIndex}-yearmonth-month`);
-        return doesAnyFieldsHaveValue() ? validateYearMonthMax(maxDateTime, getYearFromString(value), monthValue, resources) : true;
+        return doesAnyFieldsHaveValue() ? validateYearMonthMax(maxDateTime, getYearFromString(value), monthValue, resources, item) : true;
       },
     },
     shouldUnregister: true,
   };
 
   const validationRulesMonth: RegisterOptions<FieldValues, string> | undefined = {
-    required: {
-      value: isRequired(item),
-      message: resources?.yearmonth_field_required || '',
-    },
+    required: required({ item, resources, message: resources?.yearmonth_field_required }),
     shouldUnregister: true,
   };