+++ CHANGES.md
@@ -1,3 +1,8 @@
+## 20.1.0
+
+- Replaced custom validation-summary with validation-summary component from HN-Designsystem
+- Use presentationButtonsType on custom buttons
+
 ## 20.0.0
 
 - Updated peerDependencies @helsenorge packages to v.36.x.x and 12.x.x